move itineraryItem to entities
