+++ wetmap/src/entities/itineraryItem.ts
@@ -1,0 +1,13 @@
+export type ItineraryItem = {
+    BookingPage: string
+    created_at:  string
+    description: string
+    id:          number
+    name:        string
+    price:       string
+    shopID:      number
+    siteList:    number[]
+    startDate:   string
+    endDate:    string
+    tripName:    string
+  };
